Don't turn absolute time to delta time when transforming
Since we are in a data-first modeling, during fit columns
are in terms of delta-time (`t-1`, `t-2`, etc. for current
`t`).
However, during transforming, having this reverse order
is confusing
(e.g., for `T=5`, `t=1` will become `t=5-1=4`, `t=2` will
become `t=5-2=3`, etc.),
and it is better to keep track of absolute time.