Optional time-pooled modeling - initial commit
Fit a single model per-covariate over all time points.
Allows to pool more data together that might improve
precision (larger sample size), but might require to
adjust for the time-value as an additional variable
in some parametric way.

This commit is in a middle of a work in progress
(Committed on July 22nd 2024, but last edited on July 13th 2023).